Database Horizon Modification for Template Setup (Not yet finished)
