+++ Backend/enrollment/.env
@@ -1,6 +1,7 @@
-DB_HOST=briaraujvkkiast9gz9o-mysql.services.clever-cloud.com
-DB_NAME=briaraujvkkiast9gz9o
-DB_USER=uczowlxwiez7ujzc
+DB_HOST=localhost
+DB_NAME=enrollment_system
+DB_USER=root
 DB_PORT=3306
-DB_PASS=fKl3OAtNoiwOAIT7Z933
+DB_PASS=
 DB_URI=mysql://uczowlxwiez7ujzc:fKl3OAtNoiwOAIT7Z933@briaraujvkkiast9gz9o-mysql.services.clever-cloud.com:3306/briaraujvkkiast9gz9o
+SECRET_KEY=django-insecure-rc0ljpof7owdbfd2ic_6f&bq=)2p65r^8af7vyawf^3-*!(9k0
