+++ Backend/enrollment/api/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class AppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'api'
