+++ Backend/enrollment/frontend/apps.py
@@ -1,5 +1,6 @@
 from django.apps import AppConfig
 
-class AppConfig(AppConfig):
+
+class FrontendConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'app'
+    name = 'frontend'
