--- README.md
@@ -1,1 +1,0 @@
-Don't mind yung appinstallation. Remove ko yan next time.
